+++ robot/proto/robot.proto
@@ -38,8 +38,8 @@ enum OutPort {
 
 // Functions that can be called through the interface
 service Motors {
-  rpc StopMotors(MultipleMotors) returns (StatusReply) {}
   rpc Drive(DriveRequest) returns (StatusReply) {}
+  rpc StopMotors(MultipleMotors) returns (StatusReply) {}
   rpc RotateWGyro(RotateRequest) returns (StatusReply) {}
   rpc CollectRelease(MultipleMotors) returns (StatusReply) {}
 }
